backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
alembic
requests
